[Interfaces] JSON-RPC, make StateControl a common API
