+++ Source/interfaces/json/WebKitBrowserAPI.json
@@ -9,6 +9,11 @@
   "common": {
     "$ref": "common.json"
   },
+  "include": {
+    "statecontrol": {
+      "$ref": "StateControlAPI.json#"
+    }
+  },
   "methods": {
     "WebKitBrowser.1.status": {
       "summary": "Retrieves the WebKit Browser information.",
@@ -19,22 +24,22 @@
           "url": {
             "description": "The currently loaded URL in the browser",
             "type": "string",
-            "default": "https://www.google.com"
+            "example": "https://www.google.com"
           },
           "fps": {
             "description": "The current number of frames per second the browser is rendering",
             "type": "number",
-            "default": 30
+            "example": 30
           },
           "suspended": {
             "description": "Determines if the browser is in suspended mode (true) or resumed mode (false)",
             "type": "boolean",
-            "default": false
+            "example": false
           },
           "hidden": {
             "description": "Determines if the browser is hidden (true) or visible (false)",
             "type": "boolean",
-            "default": false
+            "example": false
           }
         },
         "required": [
@@ -43,26 +48,6 @@
           "suspended",
           "hidden"
         ]
-      }
-    },
-    "WebKitBrowser.1.suspend": {
-      "summary": "Suspends the WebKit Browser.",
-      "description": "With this method the WebKit Browser can be suspended. Suspending when already in suspended mode has no effect.",
-      "events": [
-        "statechange"
-      ],
-      "result": {
-        "$ref": "#/common/results/void"
-      }
-    },
-    "WebKitBrowser.1.resume": {
-      "summary": "Resumes the WebKit Browser.",
-      "description": "With this method the WebKit Browser can be resumed from suspended mode. Resuming from a not suspended mode has no effect.",
-      "events": [
-        "statechange"
-      ],
-      "result": {
-        "$ref": "#/common/results/void"
       }
     },
     "WebKitBrowser.1.hide": {
@@ -97,12 +82,9 @@
           "url": {
             "description": "The URL to load",
             "type": "string",
-            "default": "https://www.google.com"
+            "example": "https://www.google.com"
           }
-        },
-        "required": [
-          "url"
-        ]
+        }
       },
       "result": {
         "$ref": "#/common/results/void"
@@ -123,32 +105,18 @@
           "url": {
             "description": "The URL that has been loaded or requested",
             "type": "string",
-            "default": "https://www.google.com"
+            "example": "https://www.google.com"
           },
           "loaded": {
             "description": "Determines if the URL has just been loaded (true) or if URL change has been requested (false)",
             "type": "boolean",
-            "default": false
+            "default": true,
+            "example": false
           }
         },
         "required": [
-          "url"
-        ]
-      }
-    },
-    "statechange": {
-      "summary": "Signals a state change in the browser.",
-      "params": {
-        "type": "object",
-        "properties": {
-          "suspended": {
-            "description": "Determines if the browser has reached suspended state (true) or resumed state (false)",
-            "type": "boolean",
-            "default": false
-          }
-        },
-        "required": [
-          "suspended"
+          "url",
+          "loaded"
         ]
       }
     },
@@ -160,12 +128,9 @@
           "hidden": {
             "description": "Determines if the browser has been hidden (true) or made visible (false)",
             "type": "boolean",
-            "default": false
+            "example": false
           }
-        },
-        "required": [
-          "hidden"
-        ]
+        }
       }
     },
     "pageclosure": {
